break up start of status phase into rearrangeable pieces
